TIP-1287: move subscribers
LocalizableSubscriber
CheckChannelsOnDeletionSubscriber
UpdateIndexesOnCategoryDeletion
ConfigureCategoryTreeForExportJobsAfterChangingTheChannelCategoryTree
