+++ src/Akeneo/Pim/Enrichment/Bundle/Resources/config/event_subscribers.yml
@@ -1,8 +1,6 @@
 parameters:
     pim_catalog.event_subscriber.create_attribute_requirement.class: Akeneo\Pim\Structure\Bundle\EventSubscriber\CreateAttributeRequirementSubscriber
-    pim_catalog.event_subscriber.localizable.class: Akeneo\Pim\Enrichment\Bundle\EventSubscriber\LocalizableSubscriber
     pim_catalog.event_subscriber.timestampable.class: Akeneo\Pim\Enrichment\Bundle\EventSubscriber\TimestampableSubscriber
-    pim_catalog.event_subscriber.category.check_channels.class: Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Category\CheckChannelsOnDeletionSubscriber
     pim_catalog.event_subscriber.reset_unique_validation.class: Akeneo\Pim\Enrichment\Bundle\EventSubscriber\ResetUniqueValidationSubscriber
     pim_catalog.event_subscriber.index_products.class: Akeneo\Pim\Enrichment\Bundle\EventSubscriber\IndexProductsSubscriber
 
@@ -16,7 +14,7 @@ services:
             - { name: doctrine.event_subscriber }
 
     pim_catalog.event_subscriber.localizable:
-        class: '%pim_catalog.event_subscriber.localizable.class%'
+        class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\AttributeOption\LocalizableSubscriber'
         arguments:
             - '@pim_catalog.context.catalog'
         tags:
@@ -28,7 +26,7 @@ services:
             - { name: doctrine.event_subscriber }
 
     pim_catalog.event_subscriber.category.check_channels:
-        class: '%pim_catalog.event_subscriber.category.check_channels.class%'
+        class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Category\OnDelete\CheckChannelsOnDeletionSubscriber'
         arguments:
             - '@translator'
         tags:
@@ -78,7 +76,7 @@ services:
             - { name: kernel.event_subscriber }
 
     pim_catalog.event_subscriber.category.update_indexes_on_category_deletion:
-        class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\UpdateIndexesOnCategoryDeletion'
+        class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Category\OnDelete\UpdateIndexesOnCategoryDeletion'
         arguments:
             - '@pim_catalog.query.get_descendent_category_codes'
             - '@akeneo_elasticsearch.client.product_and_product_model'
@@ -86,7 +84,7 @@ services:
             - { name: kernel.event_subscriber }
 
     pim_catalog.event_subscriber.update_exports_on_channel_update:
-        class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Category\ConfigureCategoryTreeForExportJobsAfterChangingTheChannelCategoryTree'
+        class: 'Akeneo\Pim\Enrichment\Bundle\EventSubscriber\Category\OnSave\ConfigureCategoryTreeForExportJobsAfterChangingTheChannelCategoryTree'
         arguments:
             - '@akeneo_batch.job.job_instance_repository'
             - '@akeneo_batch.updater.job_instance'
